Improved logging and exception handling
Returning a ReplyQueueItem from the async message loop, and then enqueuing that return. This is cleaner because all the of the ReplyQueuer operations are happening in one spot.  It's also better async code because it helps to prevent deadlocks by not potentially modifying a static object at the same time in the two different locations (even though that wasn't possible to happen in the first place, but I still think it's best practices to do it this way).